Refactor testing_utils
