+++ pallets/stellar-relay/Cargo.toml
@@ -30,13 +30,7 @@ substrate-stellar-sdk = { git = "https://github.com/pendulum-chain/substrate-ste
 
 sha2 = { version = "0.10.6", default-features = false }
 
-sp-core = { default-features = false, version = "6.0.0", git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.24", optional = true }
-sp-io = { default-features = false, version = "6.0.0", git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.24", optional = true }
-sp-runtime = { default-features = false, version = "6.0.0", git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.24", optional = true }
-orml-tokens = { git = "https://github.com/open-web3-stack/open-runtime-module-library", branch = "polkadot-v0.9.24", optional = true }
-orml-traits = { git = "https://github.com/open-web3-stack/open-runtime-module-library", branch = "polkadot-v0.9.24", optional = true }
 primitives = { package = "spacewalk-primitives", path = "../../primitives", optional = true }
-rand = { version = "0.8.5", optional = true }
 
 [dev-dependencies]
 sp-core = { default-features = false, version = "6.0.0", git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.24" }
@@ -65,7 +59,7 @@ std = [
     "substrate-stellar-sdk/std",
     "sha2/std"
 ]
-testing-utils = ["orml-tokens", "orml-traits", "primitives", "sp-core", "sp-io", "sp-runtime", "rand"]
+testing-utils = ["primitives"]
 
 runtime-benchmarks = ["frame-benchmarking/runtime-benchmarks"]
 try-runtime = ["frame-support/try-runtime"]
